+++ CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <hr/>
 
+## v1.4.3 (2018-07-27)
+
+### Added
+* [AppCommon] `/generator/database` route to generate all tables, pk indexes and views to database.
+* [AppCommon] `InstallModule` to support database creation and data population.
+* [AppCommon] `MoInstallSQL.createPKSQL()` and `MoInstallSQL.createViewSQL()` methods.
+* [AppCommon] `ConnectionService.runSyncSql()` method to run sql one after another using callbacks.
+
 ## v1.4.2 (2018-07-26)
 
 ### Added